--- TodoList/Application/Dtos/FutureDateAttribute.cs
@@ -1,13 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace TodoList.Application.Dtos;
-
-public class FutureDateAttribute: ValidationAttribute
-{
-    public override bool IsValid(object value)
-    {
-        if (value is null) return true;
-        return ((DateTime)value) > DateTime.Now;
-    }
-    
-}
