--- TodoList/Domain/Entities/Priority.cs
@@ -1,9 +1,0 @@
-namespace TodoList.Entities;
-
-public enum Priority
-{
-    LOW,
-    MEDIUM,
-    HIGH,
-    URGENT
-}
