--- TodoList/Domain/Model/Priority.cs
@@ -1,9 +1,0 @@
-namespace TodoList.Models;
-
-public enum Priority
-{
-    LOW,
-    MEDIUM,
-    HIGH,
-    URGENT
-}
