--- TodoList/Domain/Ports/In/TaskListService.cs
@@ -1,12 +1,0 @@
-using TodoList.Entities;
-
-namespace TodoList.Domain.Ports.In;
-
-public interface TaskListService
-{
-    TaskList FetchOne(int id);
-    List<TaskList> FetchAll();
-    void Create(TaskList taskList);
-    void Update(int id, TaskList taskList);
-    void Delete(int id);
-}
