--- TodoList/Program.cs
@@ -1,9 +1,0 @@
-﻿namespace TodoList;
-
-class Program
-{
-    static void Main(string[] args)
-    {
-        
-    }
-}
