+++ TodoListAPI/Domain/Entities/Priority.cs
@@ -1,0 +1,9 @@
+namespace TodoListAPI.Entities;
+
+public enum Priority
+{
+    LOW,
+    MEDIUM,
+    HIGH,
+    URGENT
+}
