+++ TodoListAPI/Domain/Model/Priority.cs
@@ -1,0 +1,9 @@
+namespace TodoListAPI.Models;
+
+public enum Priority
+{
+    LOW,
+    MEDIUM,
+    HIGH,
+    URGENT
+}
